+++ example/ios/Runner/GeneratedPluginRegistrant.m
@@ -1,0 +1,19 @@
+//
+//  Generated file. Do not edit.
+//
+
+#import "GeneratedPluginRegistrant.h"
+
+#if __has_include(<userleap_flutter/UserleapFlutterPlugin.h>)
+#import <userleap_flutter/UserleapFlutterPlugin.h>
+#else
+@import userleap_flutter;
+#endif
+
+@implementation GeneratedPluginRegistrant
+
++ (void)registerWithRegistry:(NSObject<FlutterPluginRegistry>*)registry {
+  [UserleapFlutterPlugin registerWithRegistrar:[registry registrarForPlugin:@"UserleapFlutterPlugin"]];
+}
+
+@end